prevent_alian_invasion/requirements.txt
asgiref==3.5.1
CacheControl==0.12.11
cachy==0.3.0
certifi==2021.10.8
charset-normalizer==2.0.12
cleo==0.8.1
clikit==0.6.2
crashtest==0.3.1
distlib==0.3.4
Django==3.2.12
djangorestframework==3.13.1
filelock==3.6.0
html5lib==1.1
idna==3.3
keyring==21.8.0
lockfile==0.12.2
msgpack==1.0.3
packaging==20.9
pastel==0.2.1
pexpect==4.8.0
pkginfo==1.8.2
platformdirs==2.5.2
poetry==1.1.12
poetry-core==1.0.8
ptyprocess==0.7.0
pylev==1.4.0
pyparsing==3.0.8
python-dotenv==0.20.0
pytz==2022.1
requests==2.27.1
requests-toolbelt==0.9.1
shellingham==1.4.0
six==1.16.0
sqlparse==0.4.2
tomlkit==0.10.2
urllib3==1.26.9
virtualenv==20.14.1
webencodings==0.5.1
